Enable loading of refernces schemas from local file system.
When a schema B is loaded from the local file system via a relative
reference from a raml file, this patch will resolve the file from the
local file system if both files happen to have relative id from the same
domain.

Example:

A Raml file loaded from the local file system includes a json-schema from:
  Resources: !include schemas/oic.wk.res-schema.json
The 'oic.wk.res-schema.json' has an id:
  "id": "http://openconnectivityfoundation.org/core/schemas/oic.wk.res-schema.json#",
and refernences to another schema:
    "$ref": "oic.oic-link-schema.json#"

Because the referncing schema 'oic.wk.res-schema.json'
was loaded from the local file system,
and the referenced schema is in a location relative to the referenced
schema, the file will be loaded from the file system too, relative to
the referncing schema. In this simple case, even from the same folder.