Enable loading of referenced schemas with FQN from local file system.
When the following conditions are met:

1) Schema A and Schema B both have id's with absolute URL's but happen
to be in the same domain
2) And schema A was loaded from the local file system.

then

Schema B will be loaded from the local file system based on the relative
path between the two.

This allows developing dependent schema sets locally without having to
upstream them first (and still allow to validate the schemas as if they
were hosted on a web server).

Example:

A Raml file loaded from the local file system includes a json-schema from:
  Resources: !include schemas/oic.wk.res-schema.json
The 'oic.wk.res-schema.json' has an id:
  "id": "http://openconnectivityfoundation.org/core/schemas/oic.wk.res-schema.json#",
and refernences to another schema:
    "$ref": "oic.oic-link-schema.json#"

Because the referencing schema 'oic.wk.res-schema.json'
was loaded from the local file system,
and the referenced schema is in a location relative to the referenced
schema, the file will be loaded from the file system too, relative to
the referencing schema. In this simple case, even from the same folder.